test(stella-tools): split search/engine/tests.rs by subject
crates/stella-tools/src/search/engine/tests.rs was 1489 lines after #4456,
one line from the ungrandfathered ceiling (#4494). It already used the
sibling-tests.rs pattern for budget/ceiling/refresh/shims/visibility; this
extends it to the remaining #[test]s, grouped by subject the same way
[topic] naming already worked for the existing siblings. The shared fixture
(FIXTURE, ConceptEmbedder/BrokenEmbedder, indexed_fixture/embedded_fixture,
content_of) stays in the thin parent since every new topic needs it.

Two call sites used `super::X` to reach the parent's own `use super::{...}`
imports (merge_notes, warm_chunk_vectors_with_progress, coverage_note,
semantic_hits); moved into the new modules, `super::` there means `tests`
instead of `engine`, so those four names are added to the parent's existing
`use super::{...}` list and the call sites drop the now-redundant `super::`
prefix.

Every test name is unchanged, only relocated:
  ladder.rs:
    one_search_call_answers_what_today_takes_several_tools
    a_surface_embedder_cannot_answer_the_same_question
    without_an_embedder_the_answer_is_useful_and_says_it_is_a_name_match
    with_no_index_at_all_the_file_scan_answers_and_labels_itself
    each_depth_renders_a_superset_of_the_one_below
    the_default_answer_carries_the_body_without_a_follow_up_read
  disclosure.rs:
    a_capped_scan_discloses_what_it_never_saw
    a_capped_scan_sees_the_root_before_the_leaves
    a_cut_name_match_list_discloses_how_many_matched
    a_misconfigured_embedder_note_joins_the_cut_list_note_instead_of_replacing_it
    merging_notes_loses_neither
    a_doc_comment_walk_skips_attributes_instead_of_quoting_them
    a_truncated_answer_says_it_was_truncated
    the_rendered_hit_count_matches_the_allocation
  ranking.rs:
    an_exact_symbol_name_outranks_the_semantic_guess
    the_matched_symbol_leads_the_detailed_facets
    caller_entries_carry_their_site_not_a_placeholder
    the_report_preserves_rank_order_and_strategy_labels
    an_exact_symbol_name_beats_the_ranking_that_would_bury_it
    an_exact_name_without_an_embedder_is_still_a_certainty
    a_sentence_never_reaches_the_exact_symbol_rung
  eager_pass.rs:
    one_eager_pass_embeds_every_pending_chunk_no_matter_how_many_files
    an_eager_chunk_pass_reports_progress_as_files_commit
    a_capped_pass_reports_what_it_left_pending
    a_file_whose_symbols_collide_on_rendered_text_does_not_spin_the_pass
    a_fully_embedded_workspace_stops_calling_itself_partial
    a_workspace_with_work_left_still_says_it_is_partial
    an_eager_pass_embeds_the_corpus_and_is_idempotent
    an_eager_pass_reports_progress_per_batch
    a_pass_that_hits_its_cap_reports_what_it_left
    an_unreadable_file_does_not_stop_the_eager_pass_at_the_readable_ones
    a_broken_backend_makes_the_eager_pass_a_named_failure

No baseline entry added or needed post-split.

Refs #4494
